test(analytics): organise the suite by subject, not by what prompted it
`test_review_fixes.py` was named after a review round. Once this merges
that context is gone and the name says nothing about what is covered, so
its tests move to the files that own their subject:

  runtime separator collapsing  -> test_event_names.py
  queue-full claim release      -> test_api.py
  non-string action             -> test_api.py
  rule without __name__         -> test_rules.py
  session properties after fork -> test_fork.py

Two of them had no subject file, so there are two new ones:

  test_lifecycle.py  starting and stopping the machinery - refusing to
                     start with no destination, and not starting after
                     a shutdown has already landed
  test_sender.py     getting events onto the wire - which answers it
                     carries past, which one ends reporting, and TLS

`test_rejection.py` keeps only the reaction to a rejection, its own
subject; the two status tests that decide what counts as one belong with
the sender. The event/sender builders both files were duplicating are now
conftest factories.

Splitting the file surfaced an order dependency that was already there:
the end-to-end rejection test passed in file order and failed alone. It
replaces `OpikConfig` for the whole SDK with a stub carrying only the two
fields `_start_worker` reads, and the worker thread reads it too when
building session properties - `AttributeError: 'C' object has no
attribute 'url_override'`, swallowed, so the event silently vanished. It
only passed because earlier tests had warmed the cache that would have
re-derived it. Both files now override a real config instead. Every file
in the package passes on its own now, which none of this did before.
